@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "20": "^1.0.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@fortawesome/fontawesome-svg-core": "^6.4.0",
@@ -28,14 +27,12 @@
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "framer-motion": "^12.6.3",
-    "install": "^0.13.0",
     "js-sha256": "^0.11.0",
     "lucide-react": "^0.488.0",
     "moment": "^2.30.1",
     "next": "^15.2.4",
     "next-secure-headers": "^2.2.0",
     "next-themes": "^0.4.6",
-    "npm": "^11.3.0",
     "react": "^18.2.0",
     "react-cookie-consent": "^8.0.1",
     "react-copy-to-clipboard": "^5.1.0",
@@ -59,11 +56,8 @@
     "simplebar-react": "^3.3.2",
     "sweetalert2": "^11.6.13",
     "swr": "^1.3.0",
-    "tailwind": "^2.3.1",
     "tailwind-merge": "^3.3.1",
-    "tailwindcss-elevation": "^2.0.0",
-    "uuidv4": "^6.2.13",
-    "yarn": "^1.22.19",
+    "uuid": "^9.0.0",
     "yup": "^0.32.11",
     "zustand": "^5.0.6"
   },
@@ -76,8 +70,8 @@
     "@types/react-helmet": "^6.1.6",
     "@types/react-textfit": "^1.1.4",
     "autoprefixer": "^10.4.21",
-    "eslint": "^7.32.0",
-    "eslint-config-next": "^12.0.4",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^15.2.4",
     "js-cookie": "^3.0.5",
     "postcss": "^8.5.6",
     "tailwindcss": "^3.4.17",
